+++ test/cmdlineTests/event_from_another_contract_userdoc_devdoc/input.sol
@@ -1,0 +1,11 @@
+contract X {
+    /// @notice E event
+    /// @dev E event
+    event E();
+}
+
+contract C {
+    function g() public {
+        emit X.E();
+    }
+}